requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
python-dotenv
email_validator
